Temp version

+++ benchmarks/tpch_query.cpp
@@ -23,13 +23,18 @@ void tpch_query::gen_Q6(uint64_t thd_id) {
 	part_to_access[0] = part_id;
 	part_num = 1;
 
+
 	//Q6 related vars
-	// from 93~97,+(1,365)
 	uint64_t year = URand(93, 97, 0);
 	uint64_t day = URand(1, 365, 0);
 	date = (uint64_t)(year * 1000 + day);
 	discount = ((double)URand(2, 9, 0)) / 100;
-	quantity = URand(24, 25, 0);
+	quantity = (double)URand(24, 25, 0);
+
+	// // test
+	// date = (uint64_t)95222;
+	// discount = (double)0.06;
+	// quantity = double(24);
 
 
 }
